+++ packages/gamut/src/GridForm/GridFormInputGroup/GridFormCheckboxInput/index.tsx
@@ -21,6 +21,7 @@ export const GridFormCheckboxInput: React.FC<GridFormCheckboxInputProps> = ({
       defaultChecked={field.defaultValue}
       htmlFor={field.name}
       name={field.name}
+      onChange={event => field.onUpdate?.(event.target.checked)}
       label={field.description}
       ref={register(field.validation)}
     />