+++ packages/gamut/src/GridForm/GridFormInputGroup/GridFormFileInput/index.tsx
@@ -23,6 +23,7 @@ export const GridFormFileInput: React.FC<GridFormFileInputProps> = ({
       error={error}
       htmlFor={field.name}
       name={field.name}
+      onChange={event => field.onUpdate?.(event.target.files)}
       ref={register(field.validation)}
       type="file"
     />