+++ packages/styleguide/.storybook/decorators/wrapper.js
@@ -1,9 +1,2 @@
-import React from 'react';
 import '../../../gamut-styles/core/index';
 import './styles';
-
-export const wrapper = story => {
-  return story();
-};
-
-export default wrapper;