+++ packages/styleguide/.storybook/preview.js
@@ -1,10 +1,6 @@
-import { addParameters, configure, addDecorator } from '@storybook/react';
+import { addParameters } from '@storybook/react';
 import { create } from '@storybook/theming';
-import { withA11y } from '@storybook/addon-a11y';
-import wrapper from './decorators/wrapper';
-
-addDecorator(withA11y);
-addDecorator(wrapper);
+import './decorators/wrapper';
 
 const gamutTheme = create({
   base: 'light',
@@ -31,5 +27,11 @@ addParameters({
       method: 'alphabetical',
       locales: 'en-US',
     },
+  },
+  a11y: {
+    element: '#root',
+    config: {},
+    options: {},
+    manual: true,
   },
 });