+++ packages/webpack-config/src/config/css-extracted.js
@@ -5,6 +5,7 @@ const loaders = require('../loaders');
 const cssExtracted = options => {
   const defaultExtractOptions = {
     filename: '[name].css',
+    esModule: true,
     ignoreOrder: true,
   };