+++ packages/styleguide/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 9.0.0 (2020-04-28)
+
+* feat(webpack-config)!: Update CSS Loader and move loaders into dependencies (#736) ([2e6c1c5](https://github.com/RyzacInc/client-modules/commit/2e6c1c5)), closes [#736](https://github.com/RyzacInc/client-modules/issues/736)
+
+
+
+
+
+## 8.5.0 (2020-04-28)
+
+* feat(styleguide): added inclusion guidelines to FAQs (#749) ([69d2a40](https://github.com/RyzacInc/client-modules/commit/69d2a40)), closes [#749](https://github.com/RyzacInc/client-modules/issues/749)
+
+
+
+
+
+## 8.4.0 (2020-04-28)
+
+* feat(Gamut): switched ProgressBar back to taking in a 'style' prop (#752) ([eb4f640](https://github.com/RyzacInc/client-modules/commit/eb4f640)), closes [#752](https://github.com/RyzacInc/client-modules/issues/752)
+
+
+
+
+
+## 8.3.0 (2020-04-27)
+
+* feat(grid-form): add onUpdate to GridForm fields (#751) ([b465604](https://github.com/RyzacInc/client-modules/commit/b465604)), closes [#751](https://github.com/RyzacInc/client-modules/issues/751)
+
+
+
+
+
 ## 8.2.0 (2020-04-22)
 
 * feat(Gamut): added ProgressBar minimumPercent and theme props (#739) ([b17a7bc](https://github.com/RyzacInc/client-modules/commit/b17a7bc)), closes [#739](https://github.com/RyzacInc/client-modules/issues/739)